+++ Sources/Rhythm/RhythmTree.swift
@@ -12,7 +12,7 @@ import MetricalDuration
 
 public struct Rhythm <T> {
     public let metricalDurationTree: MetricalDurationTree
-    public let leaves: [RhythmLeaf<T>]
+    public let leaves: [Leaf]
 }
 
 extension Rhythm {
@@ -27,13 +27,9 @@ extension Rhythm {
 
 extension Rhythm {
 
-    public init(
-        _ metricalDurationTree: MetricalDurationTree,
-        _ leafContexts: [MetricalContext<T>]
-    )
-    {
+    public init(_ metricalDurationTree: MetricalDurationTree, _ leaves: [MetricalContext<T>]) {
         self.metricalDurationTree = metricalDurationTree
-        self.leaves = zip(metricalDurationTree.leaves, leafContexts).map(RhythmLeaf.init)
+        self.leaves = zip(metricalDurationTree.leaves, leaves).map(Leaf.init)
     }
 }
 
@@ -41,7 +37,7 @@ public func lengths <S,T> (of rhythmTrees: S) -> [MetricalDuration]
     where S: Sequence, S.Element == Rhythm<T>
 {
     func merge(
-        _ leaves: ArraySlice<RhythmLeaf<T>>,
+        _ leaves: ArraySlice<Rhythm<T>.Leaf>,
         into accum: [MetricalDuration],
         tied: MetricalDuration?
     ) -> [MetricalDuration]
@@ -72,6 +68,38 @@ public func lengths <S,T> (of rhythmTrees: S) -> [MetricalDuration]
 
     return merge(ArraySlice(rhythmTrees.flatMap { $0.leaves }), into: [], tied: nil)
 }
+
+extension Rhythm {
+
+    /// Leaf item of a hierarchically-structured `Rhythm`.
+    public struct Leaf {
+
+        public let metricalDuration: MetricalDuration
+        public let context: MetricalContext<T>
+
+        public init(metricalDuration: MetricalDuration, context: MetricalContext<T>) {
+            self.metricalDuration = metricalDuration
+            self.context = context
+        }
+
+        public func map <U> (_ transform: @escaping (T) -> U) -> Rhythm<U>.Leaf {
+
+            // FIXME: Extract this into func. Generics not happy.
+            var newContext: MetricalContext<U> {
+                switch context {
+                case .continuation:
+                    return .continuation
+                case .instance(let instance):
+                    return .instance(instance.map(transform))
+                }
+            }
+
+            return Rhythm<U>.Leaf(metricalDuration: metricalDuration, context: newContext)
+        }
+    }
+}
+
+extension Rhythm.Leaf: Equatable where T: Equatable { }
 
 /// - returns: `RhythmTree` with the given `MetricalDurationTree` and `MetricalContext` values.
 public func * <T> (lhs: MetricalDurationTree, rhs: [MetricalContext<T>]) -> Rhythm<T> {
